Fixed issue where previous messages from the calculation were not bei… (#64)
* Fixed issue where previous messages from the calculation were not being cleared before starting a new one.  This caused cookie overflow errors.

Also now freezing the calculation on completion.  This freezing is done firstly to prevent modification by accident - and secondly, the controller detects it and redirects the user back to the start if the current calculation is frozen

* Fixed issue where previous messages from the calculation were not being cleared before starting a new one.  This caused cookie overflow errors.

Also now freezing the calculation on completion.  This freezing is done firstly to prevent modification by accident - and secondly, the controller detects it and redirects the user back to the start if the current calculation is frozen