RST-947 - Guidance Text Link In IE and Edge
The GDS standards recommend we use details and summary tags which modern browsers support, so they provided a polyfill for older browsers.

However, this polyfill was not very well written which means it does not play well with pages that are constantly changing (such as any rails app with turbolinks enabled).
It was also hooking into all click events even ones that it should not care about and intercepting them to see if they are a summary tag that has a parent details tag !!

So, I have re written this and simply not used their code - this now works and is mainly done in CSS.  The javascript simply toggles the 'open' attribute in the details tag.

It now also supports aria as the old one did